backend/pyproject.toml
[tool.poetry]
name = "next-graphql-backend"
version = "0.1.0"
description = "Python バックエンド with Strawberry GraphQL, SQLAlchemy, FastAPI"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
strawberry-graphql = "0.196.0"
sqlalchemy = "2.0.28"
psycopg2-binary = "2.9.9"
fastapi = "0.115.0"
uvicorn = "0.27.1"
alembic = "1.12.1"
python-dotenv = "1.0.0"
anyio = "4.7.0"
starlette = "0.38.1"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 
